Translate text

Translate strings, vectors or lists in the Server of a Shiny app. There are currently 15 languages available for translation (see Available languages below). Note that the i_ function should be used within the Server of the app and the ui_ function should be used in the UI.

i_(str, lang = NULL, i18n = NULL, markdown = FALSE, keys = c("name", "label"))

ui_(string, lang = NULL)

Arguments

str
A string, vector or list to be translated.

lang
Code for the language that the original language should be translated into.

i18n
List of language configurations; can only be set for i_. TODO add default configurations! Options that can be set are:

defaultLang Default language used in Shiny app; default = "en"

availableLangs Language that can be chosen for translation in Shiny app; there are currently 15 available languages (see Available languages); defaults to all available languages

localeDir Directory to yaml files which contain custom keyword translations; default = "locale"

fallbacks List of fallback languages if translation for a word is not found in desired language; defaults to Default fallbacks

queryParameter String to define query parameter if language to be set through URL; default = "lang"

markdown
TODO; can only be set for i_.

keys
If str is a list this is a string (or a vector of strings) specifying which key(s) of the list to translate; can only be set for i_.

Value

Translation of input text in the same format as the input.

Available languages

There are currently 15 languages available for translation:

code	language

ar	Arabic

ca	Catalan

da	Danish

de	German

en	English

es	Spanish

fr	French

he	Hebrew

hi	Hindi

it	Italian

pt	Portuguese

pt_BR	Portuguese (Brazil)

ru	Russian

sv	Swedish

zh_CN	Chinese

Default fallbacks

If no fallback languages are specified, translations automatically fall back onto the following languages.

original language	fallback language

es	pt

pt	es

fr	pt

de	nl

nl	de

Examples

i_("hello", lang = "de")

#> locale directory does not exists, using default translations

#> [1] "Hallo"

i_(c("hello", "world"), lang = "de")

#> locale directory does not exists, using default translations

#> [1] "Hallo" NA     

i_(list(id = "hello", translate = "world"), lang = "es", keys = "translate")

#> locale directory does not exists, using default translations

#> $id
#> [1] "hello"
#> 
#> $translate
#> [1] "mundo"
#>